+++ starcitizen_xml_editor.ps1
@@ -9,7 +9,7 @@
               ███    ███  The VRse Attribute Editor  Author: @troubleshooternz
 #>
 
-$scriptVersion = "0.2"                        # fov wizard version 0.2 has dependencies for both powershell and python and must be run from a command prompt, not powershell. bat file included in the repo.
+$scriptVersion = "0.2.1"                        # 0.2.1 batch file launching stuff, moved some buttons to the Actions Menu, auto populate form on start up, more descriptive tool tips
 $BackupFolderName = "VRSE AE Backup"
 $profileContent = @()
 $script:profileArray = [System.Collections.ArrayList]@()
@@ -592,7 +592,7 @@ $saveProfileMenuItem.Add_Click({
 })
 $fileMenuItem.MenuItems.Add($saveProfileMenuItem)  # Add the Save Profile menu item to the File menu
 
-# Add an item to the menu called "Open XML"
+
 $actionsMenuItem = New-Object System.Windows.Forms.MenuItem
 $actionsMenuItem.Text = "&Actions"
 $mainMenu.MenuItems.Add($actionsMenuItem)
@@ -601,9 +601,18 @@ $helpMenuItem = New-Object System.Windows.Forms.MenuItem
 $helpMenuItem.Text = "&Help"
 $mainMenu.MenuItems.Add($helpMenuItem)
 
+$loadsettingsfromGameMenuItem = New-Object System.Windows.Forms.MenuItem
+$loadsettingsfromGameMenuItem.Text = "&Load Settings from Game"
+$loadsettingsfromGameMenuItem.Add_Click({
+    Import-SettingsFromGame
+})
+$actionsMenuItem.MenuItems.Add($loadsettingsfromGameMenuItem)  # Add the Load Settings from Game menu item to the Actions menu
 
+# Add an item to the menu called "Open XML"
 $openXmlMenuItem = New-Object System.Windows.Forms.MenuItem
 $openXmlMenuItem.Text = "&Open XML"
+$openXmlMenuItem.Enabled = $false
+$openXmlMenuItem.Visible = $false
 
 $openXmlMenuItem.Add_Click({
     $openFileDialog = New-Object System.Windows.Forms.OpenFileDialog
@@ -1156,17 +1165,7 @@ function SetComboBoxValue {
     }
 }
 
-
-$importButton = New-Object System.Windows.Forms.Button
-$importButton.Text = "Import settings from Game"
-$importButton.Name = "ImportButton"
-$importButton.Width = 200
-$importButton.Height = 30
-$importButton.Top = 30
-$importButton.Left = 260
-$importButton.TabIndex = 5
-
-$importButton.Add_Click({
+function Import-SettingsFromGame {
     try {
         $script:xmlContent = [xml](Get-Content $script:xmlPath)
         if ($script:xmlContent.DocumentElement.ChildNodes.Count -gt 0) {
@@ -1206,6 +1205,23 @@ $importButton.Add_Click({
         Set-ProfileArray
         }
     }
+}
+
+$importButton = New-Object System.Windows.Forms.Button
+$importButton.Text = "Import settings from Game"
+$importButton.Name = "ImportButton"
+$importButton.Width = 200
+$importButton.Height = 30
+$importButton.Top = 30
+$importButton.Left = 260
+$importButton.TabIndex = 5
+$importButton.TabStop = $false
+$importButton.Visible = $false  # Initially hidden
+$importButton.Enabled = $false  # Initially disabled
+
+
+$importButton.Add_Click({
+    Import-SettingsFromGame
 })
 
 # Initially disable the import and save buttons
@@ -1694,10 +1710,22 @@ $saveButton.Add_Click({
             $HeadtrackingThirdPersonCameraToggleNode.SetAttribute("value", $HeadtrackingThirdPersonCameraToggleComboBox.SelectedIndex.ToString())  # HEADTRACKINGTHIRDPERSONCAMERATOGGLE
         }
         # Save the XML content to the specified path
+        # Check if VorpX task is running
+
+        $vorpxRunning = Get-Process -Name "vorpControl" -ErrorAction SilentlyContinue
+        if ($vorpxRunning) {
+            #$statusBar.Text = "VorpX is running."
+            $vorpxindicatorText = "VorpX is running."
+        } else {
+            #$statusBar.Text = "VorpX is NOT running."
+            $vorpxindicatorText = "VorpX is NOT running."
+        }
 
         try {
             $script:xmlContent.Save($script:xmlPath)
-            [System.Windows.Forms.MessageBox]::Show("Values saved successfully!")
+            [System.Windows.Forms.MessageBox]::Show("Saved. You may now close this window. Remember to start VorpX Control Panel and start the listener before launching Star Citizen!" + 
+            "`n`n" +
+            $vorpxindicatorText)
         } catch {
             [System.Windows.Forms.MessageBox]::Show("An error occurred while saving the XML file to $script:xmlPath: $_")
         }
@@ -1840,20 +1868,20 @@ $ShowHelp={
         "heightTextBox" {$tip = "Height of the screen in pixels"}
         "headtrackerEnabledComboBox" {$tip = "Enable or disable head tracking"}
         "HeadtrackingSourceComboBox" {$tip = "Select the head tracking source"}
-        "chromaticAberrationTextBox" {$tip = "Chromatic Aberration value 0.00/1.00"}
+        "chromaticAberrationTextBox" {$tip = "Chromatic Aberration value 0.00/1.00. Recommended value 0.00"}
         #"AutoZoomTextBox" {$tip = "Auto Zoom on selected target 0/1"}
-        "AutoZoomComboBox" {$tip = "Auto Zoom on selected target 0/1"}
-        "MotionBlurTextBox" {$tip = "Motion Blur value 0/1"}
-        "ShakeScaleTextBox" {$tip = "Shake Scale value"}
-        "CameraSpringMovementTextBox" {$tip = "Camera Spring Movement value 0/1"}
-        "FilmGrainTextBox" {$tip = "Film Grain value 0/1"}
-        "GForceBoostZoomScaleTextBox" {$tip = "G-Force Boost Zoom Scale value"}
-        "GForceHeadBobScaleTextBox" {$tip = "G-Force Head Bob Scale value"}
-        "HeadtrackingEnableRollFPSComboBox" {$tip = "Enable Roll FPS 0/1"}
-        "HeadtrackingDisableDuringWalkingComboBox" {$tip = "Disable Headtracking during walking 0/1"}
-        "HeadtrackingThirdPersonCameraToggleComboBox" {$tip = "Enable Headtracking in Third Person 0/1"}
-        "saveButton" {$tip = "Save settings to the game"}
-        "saveProfileButton" {$tip = "Save settings to the profile"}
+        "AutoZoomComboBox" {$tip = "Auto Zoom on selected target. Recommended Disabled"}
+        "MotionBlurTextBox" {$tip = "Motion Blur. Recommended Disabled"}
+        "ShakeScaleTextBox" {$tip = "Shake Scale value. Recommended value 0"}
+        "CameraSpringMovementTextBox" {$tip = "Camera Spring Movement value. Recommended value 0"}
+        "FilmGrainTextBox" {$tip = "Film Grain. Recommended Disabled"}
+        "GForceBoostZoomScaleTextBox" {$tip = "G-Force Boost Zoom Scale value. valid value 0.0 to 1.0. Recommended value 0."}
+        "GForceHeadBobScaleTextBox" {$tip = "G-Force Head Bob Scale value. valid value 0.0 to 1.0. Recommended value 0."}
+        "HeadtrackingEnableRollFPSComboBox" {$tip = "Sets whether head-tilt to left/right is enabled in FPS mode.May also apply in vehicles"}
+        "HeadtrackingDisableDuringWalkingComboBox" {$tip = "Disable Headtracking during walking On/Off"}
+        "HeadtrackingThirdPersonCameraToggleComboBox" {$tip = "Enable Headtracking in Third Person On/Off"}
+        "saveButton" {$tip = "Save this configuration to the game"}
+        "saveProfileButton" {$tip = "Save these settings to a config file for later use"}
         "loadFromProfileButton" {$tip = "Load settings from the VRSE-AE profile"}
         "importButton" {$tip = "Import settings from the game"}
         "deleteEACTempFilesButton" {$tip = "Delete EAC TempFiles"}
@@ -1879,6 +1907,7 @@ if (($null -ne $AutoDetectSCPath) -and (Test-Path -Path $AutoDetectSCPath)) {
     if (Test-Path -Path $AutoDetectSCPath) {
         $importButton.Enabled = $true
         $statusBar.Text = "Star Citizen found at: $script:liveFolderPath"
+        Import-SettingsFromGame
     } else {
         $statusBar.Text = "attributes.xml file not found in the 'default' profile folder."
         #[System.Windows.Forms.MessageBox]::Show("attributes.xml file not found in the 'default' profile folder.")
@@ -1967,6 +1996,7 @@ function Open-FovWizard {
 $chooseFovWizardButton = New-Object System.Windows.Forms.Button
 $chooseFovWizardButton.Name = "ChooseFovWizardButton"
 $chooseFovWizardButton.Text = "FOV Wizard"
+$chooseFovWizardButton.Font = New-Object System.Drawing.Font($chooseFovWizardButton.Font.FontFamily, $chooseFovWizardButton.Font.Size, [System.Drawing.FontStyle]::Bold)
 $chooseFovWizardButton.Width = 80
 $chooseFovWizardButton.Height = 30
 $chooseFovWizardButton.Top = 65